# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HRUO Enterprise HCM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── BRANCHES ──────────────────────────────────────────────────────────────
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List branches
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ── JOB LEVELS ────────────────────────────────────────────────────────────
  /job-levels:
    get:
      operationId: listJobLevels
      tags: [jobLevels]
      summary: List job levels
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobLevel"
    post:
      operationId: createJobLevel
      tags: [jobLevels]
      summary: Create job level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobLevelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobLevel"

  # ── LEAVE ─────────────────────────────────────────────────────────────────
...
  /payroll/payslips:
    get:
      operationId: listPayslips
      tags: [payroll]
      summary: List payslips
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payslip"
    post:
      operationId: createPayslip
      tags: [payroll]
      summary: Process payroll / create payslip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayslipInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payslip"
  /payroll/payslips/{id}:
    get:
      operationId: getPayslip
      tags: [payroll]
      summary: Get payslip details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payslip"
  /payroll/tax-calculator:
    post:
      operationId: calculateTax
      tags: [payroll]
      summary: Calculate tax, insurance, end-of-service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxCalculatorInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxCalculatorResult"
  /payroll/summary:
    get:
      operationId: getPayrollSummary
      tags: [payroll]
      summary: Get payroll summary for a period
      parameters:
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollSummary"

  # ── RECRUITMENT ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Department:
      type: object
      required: [id, name, code, active]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        active:
          type: boolean
        parentId:
          type: ["integer", "null"]
        headId:
          type: ["integer", "null"]
        headName:
          type: ["string", "null"]
        employeeCount:
          type: integer
    DepartmentInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        parentId:
          type: ["integer", "null"]
        headId:
          type: ["integer", "null"]
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        active:
          type: boolean
        parentId:
          type: ["integer", "null"]
        headId:
          type: ["integer", "null"]

    # ── BRANCH ────────────────────────────────────────────────────────────────
    Branch:
      type: object
      required: [id, name, code, active]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        active:
          type: boolean
        address:
          type: ["string", "null"]
        timezone:
          type: ["string", "null"]
        employeeCount:
          type: integer
    BranchInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        address:
          type: string
        timezone:
          type: string
    BranchUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        active:
          type: boolean
        address:
          type: string
        timezone:
          type: string

    # ── JOB LEVEL ─────────────────────────────────────────────────────────────
    JobLevel:
      type: object
      required: [id, title, code, active]
      properties:
        id:
          type: integer
        title:
          type: string
        code:
          type: string
        active:
          type: boolean
        level:
          type: ["string", "null"]
    JobLevelInput:
      type: object
      required: [title, code]
      properties:
        title:
          type: string
        code:
          type: string
        level:
          type: string

    # ── LEAVE ─────────────────────────────────────────────────────────────────
...
    Payslip:
      type: object
      required: [id, employeeId, period, grossSalary, netSalary, status]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        period:
          type: string
        grossSalary:
          type: number
        netSalary:
          type: number
        totalDeductions:
          type: number
        totalEarnings:
          type: number
        socialInsurance:
          type: number
        incomeTax:
          type: number
        status:
          type: string
          enum: [draft, published]
        publishedAt:
          type: ["string", "null"]
        currency:
          type: string
    PayslipInput:
      type: object
      required: [employeeId, period]
      properties:
        employeeId:
          type: integer
        period:
          type: string
        isDraft:
          type: boolean
        bonuses:
          type: number
        deductions:
          type: number
        remarks:
          type: string
    TaxCalculatorInput:
      type: object
      required: [grossSalary, country]
      properties:
        grossSalary:
          type: number
        country:
          type: string
          enum: [egypt, tunisia, morocco, gcc, usa, europe]
        yearsOfService:
          type: number
        terminationType:
          type: string
          enum: [resignation, dismissal, retirement]
    TaxCalculatorResult:
      type: object
      required: [grossSalary, incomeTax, socialInsurance, netSalary]
      properties:
        grossSalary:
          type: number
        incomeTax:
          type: number
        incomeTaxRate:
          type: number
        socialInsurance:
          type: number
        socialInsuranceRate:
          type: number
        netSalary:
          type: number
        endOfServiceBonus:
          type: number
        country:
          type: string
        breakdown:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              amount:
                type: number
    PayrollSummary:
      type: object
      required: [period, totalGross, totalNet, totalEmployees]
      properties:
        period:
          type: string
        totalGross:
          type: number
        totalNet:
          type: number
        totalDeductions:
          type: number
        totalTax:
          type: number
        totalInsurance:
          type: number
        totalEmployees:
          type: integer
        currency:
          type: string

    # ── RECRUITMENT ───────────────────────────────────────────────────────────
...